events/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.7.1
Django==4.2.11
django-allauth==0.61.1
django-bootstrap-v5==1.0.11
django-ckeditor==6.7.1
django-environ==0.11.2
django-js-asset==2.2.0
django-quill-editor==0.1.40
django-tinymce==4.0.0
djangorestframework==3.15.1
idna==3.6
oauthlib==3.2.2
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.4.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
urllib3==2.2.1
webencodings==0.5.1
